+++ apps/workflows-dashboard/src/pages/WorkflowDefinitions/components/molecules/WorkflowDefinitionsTable/utils/merge-columns.ts
@@ -1,0 +1,11 @@
+import { InputColumn, WorkflowTableColumnDef } from '@/components/molecules/WorkflowsTable/types';
+
+export function mergeColumns<TColumnData>(
+  leftColumn: WorkflowTableColumnDef<TColumnData>,
+  rightColumn: InputColumn<TColumnData>,
+) {
+  return {
+    ...leftColumn,
+    ...rightColumn,
+  };
+}