+++ apps/kyb-app/CHANGELOG.md
@@ -1,5 +1,27 @@
 # kyb-app
 
+## 0.3.22
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/workflow-browser-sdk@0.6.22
+  - @ballerine/blocks@0.2.8
+  - @ballerine/common@0.9.14
+  - @ballerine/ui@0.5.8
+
+## 0.3.21
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/common@0.9.13
+  - @ballerine/blocks@0.2.7
+  - @ballerine/ui@0.5.7
+  - @ballerine/workflow-browser-sdk@0.6.21
+
 ## 0.3.20
 
 ### Patch Changes