+++ sdks/web-ui-sdk/CHANGELOG.md
@@ -1,5 +1,21 @@
 # web-ui-sdk
 
+## 1.5.15
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.14
+
+## 1.5.14
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/common@0.9.13
+
 ## 1.5.13
 
 ### Patch Changes